+++ content-elements/form/friendly-captcha/prototype/index.js
@@ -18,8 +18,10 @@ module.exports = (
   elementDescription = 'Eine Friendly Captcha Lizenz wird benötigt',
   sideKeyPartId = 'friendly-captcha-part-968cbea9',
   sideKeyPartLabel = 'SiteKey',
-  friendlyCaptchaPartId = 'friendly-captcha-form-field-part-5e65c6ba',
-  friendlyCaptchaLabel = 'Nicht ändern!'
+  friendlyCaptchaPartId = 'friendly-captcha-form-field-part-bolY1C',
+  friendlyCaptchaLabel = 'Die ID darf nicht angepasst werden, da ansonsten die Funktionalität des freindly Captchas verloren geht.',
+  validityMessageId = "friendly-captcha-validity-message-Uc90KX",
+  validityMessageLabel = "Validierungsnachricht"
 
 ) => cx
 .contentElement
@@ -35,4 +37,7 @@ module.exports = (
       cx.part.formField
         .withId(friendlyCaptchaPartId)
         .withLabel(friendlyCaptchaLabel),
+      cx.part.plainText
+        .withId(validityMessageId)
+        .withLabel(validityMessageLabel)
   );