+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Next
+* Add Friendly Captcha Element
 * Fix shadow issue and add new hover configuration for table
 * Dark mode activation/deactivation using properties.js value
 * Add in form-field the file type and the phone number validation